refactor: Phase 1 - Move data layer, services, and app base classes to MVVM structure
- Move data models (DataClass, User, ChatMessage) to data/model/
- Move UserPreferences to data/local/
- Move API infrastructure (RetrofitClient, YigChik*API) to data/remote/
- Move services (TibetanKeyboard, MyFirebaseMessagingService) to service/
- Move app base classes (TibetanKeyboardApp, InputMethodActivity) to app/
- Update all package declarations and imports
- Update AndroidManifest.xml for moved components
