+++ app/src/main/java/com/kharagedition/tibetankeyboard/app/TibetanKeyboardApp.kt
@@ -1,4 +1,4 @@
-package com.kharagedition.tibetankeyboard
+package com.kharagedition.tibetankeyboard.app
 
 import android.app.Application
 import android.content.SharedPreferences
