+++ app/src/main/java/com/kharagedition/tibetankeyboard/auth/AuthManager.kt
@@ -9,7 +9,7 @@ import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 import com.kharagedition.tibetankeyboard.LoginActivity
-import com.kharagedition.tibetankeyboard.UserPreferences
+import com.kharagedition.tibetankeyboard.data.local.UserPreferences
 import com.kharagedition.tibetankeyboard.subscription.RevenueCatManager
 
 /**
