+++ app/src/main/java/com/kharagedition/tibetankeyboard/chat/ChatViewModel.kt
@@ -4,6 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.kharagedition.tibetankeyboard.data.model.ChatMessage
 import kotlinx.coroutines.launch
 class ChatViewModel : ViewModel() {
