+++ app/src/main/java/com/kharagedition/tibetankeyboard/data/model/ChatMessage.kt
@@ -1,4 +1,4 @@
-package com.kharagedition.tibetankeyboard.chat
+package com.kharagedition.tibetankeyboard.data.model
 
 import java.util.Date
