+++ app/src/main/java/com/kharagedition/tibetankeyboard/data/remote/YigChikTranslateAPI.kt
@@ -1,7 +1,7 @@
-package com.kharagedition.tibetankeyboard.ai
+package com.kharagedition.tibetankeyboard.data.remote
 
-import com.kharagedition.tibetankeyboard.model.TranslationRequest
-import com.kharagedition.tibetankeyboard.model.TranslationResponse
+import com.kharagedition.tibetankeyboard.data.model.TranslationRequest
+import com.kharagedition.tibetankeyboard.data.model.TranslationResponse
 import retrofit2.http.Body
 import retrofit2.http.Headers
 import retrofit2.http.POST
