+++ app/src/main/java/com/kharagedition/tibetankeyboard/service/MyFirebaseMessagingService.kt
@@ -1,4 +1,4 @@
-package com.kharagedition.tibetankeyboard
+package com.kharagedition.tibetankeyboard.service
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
@@ -13,6 +13,7 @@ import android.util.Log
 import androidx.core.app.NotificationCompat
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
+import com.kharagedition.tibetankeyboard.R
 
 class MyFirebaseMessagingService : FirebaseMessagingService() {
