+++ app/src/main/java/com/kharagedition/tibetankeyboard/ui/HomeActivity.kt
@@ -20,7 +20,7 @@ import com.google.android.gms.ads.*
 import com.google.android.gms.ads.nativead.NativeAd
 import com.google.android.gms.ads.nativead.NativeAdOptions
 import com.google.firebase.messaging.FirebaseMessaging
-import com.kharagedition.tibetankeyboard.application.InputMethodActivity
+import com.kharagedition.tibetankeyboard.app.InputMethodActivity
 import com.kharagedition.tibetankeyboard.R
 import com.kharagedition.tibetankeyboard.SettingsActivity
 import com.kharagedition.tibetankeyboard.ads.NativeTemplateStyle
